% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ps_biomarker_model.R
\name{fit_ps_biomarker_model}
\alias{fit_ps_biomarker_model}
\title{Fit a biomarker classification model}
\usage{
fit_ps_biomarker_model(
  ps,
  group,
  method = c("rf", "logistic"),
  ntree = 500,
  seed = 1
)
}
\arguments{
\item{ps}{A \code{phyloseq} object.}

\item{group}{Name of the \code{sample_data(ps)} column to use as the outcome variable.}

\item{method}{Character. One of \code{"rf"} or \code{"logistic"}.}

\item{ntree}{Number of trees for random forest.}

\item{seed}{Random seed used for random forest.}
}
\value{
A fitted model object.
}
\description{
Fit either a random forest or logistic regression model using taxa as predictors.
}
