% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ps_importance.R
\name{plot_ps_importance}
\alias{plot_ps_importance}
\title{Plot top biomarker taxa}
\usage{
plot_ps_importance(ranked_taxa, top_n = 20)
}
\arguments{
\item{ranked_taxa}{Output from \code{rank_ps_taxa()}.}

\item{top_n}{Number of top taxa to plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot top biomarker taxa
}
