% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{psbiomarker-package}
\alias{psbiomarker}
\alias{psbiomarker-package}
\title{psbiomarker: Biomarker Discovery in Phyloseq Objects}
\description{
Preprocess phyloseq objects, fit supervised biomarker models,
rank taxa by importance, and plot top signals.
}
\author{
\strong{Maintainer}: Brianna Perino \email{briannaperino20@gmail.com}

}
\keyword{internal}
